// src/models/product.enum.ts
export enum CategoriesEnum {
  SMARTPHONES_TV_ELECTRONICS = 1,
  COMPUTERS_LAPTOPS = 2,
  HOUSEHOLD_APPLIANCES = 3,
  SPORT = 4,
  GAME_ZONE = 5
}

export enum SubcategoriesEnum {
  MOBILE_PHONES = 1,
  TV = 2,
  NOTEBOOKS = 3,
  OFFICE_EQUIPMENT = 4,
  TABLETS = 5,
  KITCHEN = 6,
  BEAUTY = 7,
  PLAYSTATION = 8,
  FISHING = 9,
  BIKES = 10
}

export enum StateEnum {
  NEW = 1,
  PRE_OWNED = 2
}

export enum SellerEnum {
  ROZETKA = 1,
  ALLO = 2,
  FOXTROT = 3,
  MOYO = 4,
  PROM = 5
}

export enum BrandsEnum {
  SONY = 1,
  PHILIPS = 2,
  TOSHIBA = 3,
  LG = 4,
  APPLE = 5,
  SAMSUNG = 6,
  ZTE = 7,
  MOTOROLA = 8,
  TECHNO = 9,
  NOKIA = 10,
  XIAOMI = 11,
  POCO = 12,
  BEKO = 13,
  GORENJE = 14,
  INDESIT = 15,
  CORRADO = 16,
  BRAIN = 17,
  FORMULA = 18
}